+++ src/main/java/onlyoffice/managers/convert/ConvertManager.java
@@ -8,5 +8,5 @@ public interface ConvertManager extends Serializable {
     public boolean isConvertable(String ext);
     public String convertsTo(String ext);
     public String getMimeType(String ext);
-    public JSONObject convert(Long attachmentId, String ext) throws Exception;
+    public JSONObject convert(Long attachmentId, String key,  String ext) throws Exception;
 }