+++ src/main/java/onlyoffice/utils/attachment/AttachmentUtil.java
@@ -13,8 +13,14 @@ public interface AttachmentUtil extends Serializable {
     public boolean checkAccess(Attachment attachment, User user, boolean forEdit);
     public void saveAttachment(Long attachmentId, InputStream attachmentData, int size, ConfluenceUser user)
             throws IOException, IllegalArgumentException;
+    public boolean checkAccessCreate(User user, Long pageId);
     public InputStream getAttachmentData(Long attachmentId);
     public String getMediaType(Long attachmentId);
     public String getFileName(Long attachmentId);
     public String getHashCode(Long attachmentId);
+    public String getAttachmentPageTitle (Long attachmentId);
+    public Long getAttachmentPageId (Long attachmentId);
+    public String getAttachmentSpaceName (Long attachmentId);
+    public String getAttachmentSpaceKey (Long attachmentId);
+    public Attachment createNewAttachment (String title, String mimeType, InputStream file, int size, Long pageId, ConfluenceUser user) throws IOException;
 }